+++ ui/dashboards/src/components/DashboardToolbar/DashboardToolbar.tsx
@@ -22,16 +22,22 @@ import { TimeRangeControls } from '../TimeRangeControls';
 
 export interface DashboardToolbarProps {
   dashboardName: string;
+  dashboardTitleComponent?: JSX.Element;
   onEditButtonClick: () => void;
   onCancelButtonClick: () => void;
 }
 
 export const DashboardToolbar = (props: DashboardToolbarProps) => {
-  const { dashboardName, onEditButtonClick, onCancelButtonClick } = props;
+  const { dashboardName, dashboardTitleComponent, onEditButtonClick, onCancelButtonClick } = props;
 
   const { isEditMode, setEditMode } = useEditMode();
   const { openAddPanelGroup, openAddPanel } = useDashboardActions();
   const isLaptopSize = useMediaQuery(useTheme().breakpoints.up('sm'));
+  const dashboardTitle = dashboardTitleComponent ? (
+    dashboardTitleComponent
+  ) : (
+    <Typography variant="h2">{dashboardName}</Typography>
+  );
 
   const onSave = () => {
     setEditMode(false);
@@ -43,7 +49,7 @@ export const DashboardToolbar = (props: DashboardToolbarProps) => {
         <Stack spacing={2}>
           <Box sx={{ backgroundColor: (theme) => theme.palette.primary.light + '20' }}>
             <Box padding={2} display="flex">
-              <Typography variant="h2">Edit {dashboardName}</Typography>
+              {dashboardTitle}
               <Stack direction="row" spacing={1} sx={{ marginLeft: 'auto' }}>
                 <Button variant="contained" onClick={onSave}>
                   Save
@@ -79,7 +85,7 @@ export const DashboardToolbar = (props: DashboardToolbarProps) => {
       ) : (
         <Stack spacing={2} padding={2}>
           <Box sx={{ display: 'flex', width: '100%' }}>
-            <Typography variant="h2">{dashboardName}</Typography>
+            {dashboardTitle}
             <Stack direction="row" spacing={2} sx={{ marginLeft: 'auto' }}>
               <TimeRangeControls />
               {isLaptopSize && (